+++ buildSrc/src/main/kotlin/Deps.kt
@@ -1,12 +1,26 @@
 package dependencies
 
 object Deps {
-    object Kotest : Group("io.kotest") {
-        val assertions = withArtifact("kotest-assertions-core-jvm", Versions.kotest)
-        val property = withArtifact("kotest-property-jvm", Versions.kotest)
-        val runner = withArtifact("kotest-runner-junit5-jvm", Versions.kotest)
+    object Coroutines : Group("org.jetbrains.kotlinx") {
+        val core = withArtifact("kotlinx-coroutines-core", Versions.coroutines)
+        val test = withArtifact("kotlinx-coroutines-test", Versions.coroutines)
     }
-    val kotlin = dependency("org.jetbrains.kotlin:kotlin-stdlib", Versions.kotlin)
+
+    object Kotest : Group("io.kotest") {
+        val assertions = withArtifact("kotest-assertions-core", Versions.kotest)
+        val property = withArtifact("kotest-property", Versions.kotest)
+        val runner = withArtifact("kotest-runner-junit5", Versions.kotest)
+    }
+
+    object Kotlin : Group("org.jetbrains.kotlin") {
+        val reflect = withArtifact("kotlin-reflect", Versions.kotlin)
+    }
+
+    object Spek : Group("org.spekframework.spek2") {
+        val dslMetadata = withArtifact("spek-dsl-metadata", Versions.spek)
+        val dslJvm = withArtifact("spek-dsl-jvm", Versions.spek)
+        val junit5 = withArtifact("spek-runner-junit5", Versions.spek)
+    }
 }
 
 object Plugins {
@@ -25,10 +39,12 @@ private fun dependency(path: String, version: String) = "$path:$version"
 
 object Versions {
     val bintray = "1.8.5"
+    val coroutines = "1.4.0"
     val dokka = "0.9.15"
     val jacoco = "0.8.5"
     val kotest = "4.2.4"
     val kotlin = "1.4.10"
     val maven = "1.4.1"
+    val spek = "2.0.13"
     val versions = "0.33.0"
 }
