+++ remotedata/build.gradle
@@ -1,6 +1,6 @@
 import dependencies.Deps
 
-apply plugin: 'kotlin'
+apply plugin: 'org.jetbrains.kotlin.multiplatform'
 apply plugin: 'org.jetbrains.dokka'
 apply plugin: 'com.github.dcendents.android-maven'
 apply plugin: 'com.jfrog.bintray'
@@ -10,15 +10,42 @@ dokka {
     outputDirectory = "$buildDir/javadoc"
 }
 
-test {
-    useJUnitPlatform()
+kotlin {
+    jvm()
+    sourceSets {
+        commonMain {
+            dependencies {
+            }
+        }
+        commonTest {
+            dependencies {
+                implementation Deps.Kotest.assertions
+                implementation Deps.Spek.dslMetadata
+            }
+        }
+        jvmMain {
+            dependencies {
+            }
+        }
+        jvmTest {
+            dependencies {
+                implementation project(":test-util")
+                implementation Deps.Spek.dslJvm
+
+                runtimeOnly Deps.Kotlin.reflect
+                runtimeOnly Deps.Spek.junit5
+
+            }
+        }
+    }
 }
 
-dependencies {
-    compile Deps.kotlin
-
-    testCompile project(":test-util")
+jvmTest {
+    useJUnitPlatform {
+        includeEngines 'spek2', 'kotest'
+    }
 }
+
 
 ext {
     bintrayRepo = 'maven'
@@ -43,5 +70,3 @@ ext {
     licenseUrl = 'https://opensource.org/licenses/MIT'
     allLicenses = ["MIT"]
 }
-
-apply from: "$rootDir/scripts/upload.gradle"
